refactor(#3038): move turn bridge background memory observation
